+++ admin-script.js
@@ -1,6 +1,6 @@
 /**
  * Soft Dynamix Page Importer - Admin JavaScript
- * Version: 2.0.5
+ * Version: 2.0.6
  */
 (function($) {
     'use strict';